+++ CHANGES.rst
@@ -4,8 +4,9 @@ Changelog
 0.54 (unreleased)
 -----------------
 
-- Nothing changed yet.
-
+- Added `content.get_vocab_value` based on `content.get_vocab` but returns
+  the values (`attr_name='token'` by default, may also be `value` or `title`).
+  [gbastien]
 
 0.53 (2022-03-17)
 -----------------